% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering.R
\name{center}
\alias{center}
\title{Matrix Centering and Re-centering Functions}
\usage{
center(x, row = FALSE)
}
\arguments{
\item{x}{A data matrix.}

\item{row}{Whether the centering also on the rows.}

\item{object}{An object returned by any of the functions listed here.}

\item{newx}{A new data matrix.}
}
\value{
A matrix with dimensions equal to \code{x} (\code{center} or
\code{get.center}) or \code{newx} (\code{recenter}) and column means
(optionally, row means and the overall mean) stored as (an) attribute(s).
}
\description{
The function allows one to calculate, in a single step, column means, row
means, and the overall mean (\code{center} and \code{get.center}) of a matrix
an return the centered matrix (\code{center}), or center a new matrix on
values calculated previouly (\code{recenter}).
}
\details{
Function \code{center} calculates the column means (optionally, the
row means and overall mean) of a data matrix and returns the latter with its
columns centered on means of 0 (optionally, with both its columns and rows
centered on means of 0). Function \code{get.center} proceeds similarly as
\code{center}, but returns the matrix unchanged. Function \code{recenter}
takes column means (optionally, row means and an overall mean) from an object
returned by \code{center} or \code{get.center} and centers the columns
(optionally the columns and rows) of a new matrix on these values.

The functions are generally intended to be used internally but are made
available to the user willing to experiment with new ideas. They are based
on C language implementations that can also be used internally in C language
computer code.
}
\examples{
##
### First example
##
## [Examples here...]

}
\references{
[To be included...]
}
\seealso{
The \code{\link{dist}-class} and associated methods.
}
\author{
Guillaume Guénard \email{guillaume.guenard@umontreal.ca}
}
