% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuclidAB.R
\name{EuclidAB}
\alias{EuclidAB}
\title{Calculation of the Euclidean Distance Among Two Sets of Points}
\usage{
EuclidAB(a, b, squared = FALSE)
}
\arguments{
\item{a}{First matrix of cartesian coordinates.}

\item{b}{Second matrix of cartesian coordinates.}

\item{squared}{Should the squared Euclidean distances be returned (default:
FALSE).}
}
\value{
An matrix with as many rows as \code{a} and as many columns as there
are rows in \code{b}.
}
\description{
Function \code{EuclidAB} calculates the Euclidean distance between two sets
of points A and B.
}
\details{
The standard \code{R} function used to calculate the Euclidean
distance, namely \code{\link{dist}}, does only allow one to calculate
pairwise distances between the rows of a single matrix of cartesian
coordinates and return a \code{\link{dist}}-class object, which is a
one-dimensional array that is meant to be interpreted as a triangular matrix.
Function \code{Euclid} allows the user to provide two data matrices \code{a}
and \code{b} and output a rectangular Euclidean distance matrix.
}
\examples{
##
### First example
##
## [Examples here...]

}
\references{
[To be included...]
}
\seealso{
The \code{\link{dist}-class} and associated methods.
}
\author{
Guillaume Guénard \email{guillaume.guenard@umontreal.ca}
}
