% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.cov.R
\name{sp.cov}
\alias{sp.cov}
\title{Calculation of Spatial Covariance Models}
\usage{
sp.cov(d, type = c("spherical", "exponential", "power", "hyperbolic",
  "superelliptic"), alpha = 1, beta = 1)
}
\arguments{
\item{d}{Pairwise distances among objects (see details).}

\item{type}{The type of covariance model (see details).}

\item{alpha}{Shape parameter of the covariance model.}

\item{beta}{Range parameter of the covariance model.}
}
\value{
An array with the same dimensions as argument \code{d}.
}
\description{
Function \code{sp.cov} carries out the calculation of the distance-based
spatial covariance models.
}
\details{
Pairwise distances among objects can be an array with arbitrary
dimensions or a \code{\link{dist}-class} object, in which case a
one-dimensional array will be returned.

Argument \code{type} can be one of "spherical" (the default), "exponential",
"power", "hyperbolic", "superelliptic", or any unambiguous abbreviation; and
correspond to the particular function used to obtain the weights as a
function of the distances. The function does not directly return a
covariance matrix but weights that are used calculate the spatial
covariances.

Distance-weight relationships are monotonically decreasing functions
associating the level of ressemblance of two locations with the distance
separating them. For all covariance models (argument \code{type} above), the
weight is 1 at \code{d=0} and 0 beyond the range (i.e., when
\code{d >= beta}). The shape of the relationship is controled by the shape
parameter given as argument \code{alpha}. That parameter controls the
gentleness of the slope of the distance-weight relationship. It is bounded
between values of 0 (the steepest slope: a Kronecker's delta). and 1 (a
straight line).
}
\examples{
##
### First example: locations spread throughout the world
##
## [Examples here...]

}
\references{
[To be included...]
}
\seealso{
The \code{\link{dist}-class} and associated methods.
}
\author{
Guillaume Guénard \email{guillaume.guenard@umontreal.ca}
}
